:sparkles: Add support for MySQL
