+++ crud_factory/_sqlalchemy.py
@@ -2,6 +2,7 @@ from typing import Any, Dict, Generic, List, Optional, Type, TypeVar, Union
 
 from fastapi.encoders import jsonable_encoder
 from pydantic import BaseModel, ValidationError
+from sqlalchemy.exc import DataError
 from sqlalchemy.orm import Session
 from sqlalchemy.orm.exc import MultipleResultsFound, NoResultFound
 
@@ -23,7 +24,8 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
         Raises:
             TypeError: if wrong column is used on `obj_in`.
-            ValidationError: if wrong value on any `obj_in` field.
+            DataError: Invalid value on PostgreSQL.
+            OperationalError: Invalid value on MySQL.
 
         Returns:
             ModelType: Database object.
@@ -33,7 +35,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             db_obj = self._model(**obj_in_data)
             db.add(db_obj)
             db.commit()
-        except (TypeError, ValidationError) as exc:
+        except (TypeError, DataError) as exc:
             raise exc
         else:
             return db_obj
