+++ pyproject.toml
@@ -32,6 +32,7 @@ python = "^3.6"
 pydantic = "^1.7.2"
 fastapi = "^0.61.2"
 SQLAlchemy = { version = "^1.3.20", optional = true }
+mysqlclient = { version = "^2.0.1", optional = true }
 
 [tool.poetry.extras]
 sqlalchemy = ["SQLAlchemy"]
@@ -44,6 +45,7 @@ isort = "^5.6.4"
 mypy = "^0.790"
 sqlalchemy-stubs = "^0.3"
 psycopg2-binary = "^2.8.6"
+pytest-coverage = "^0.0"
 
 [tool.poetry.scripts]
 check_format = "scripts:check_format"
