+++ tests/test_sqlalchemy_factory.py
@@ -1,60 +1,9 @@
 import pytest
-from pydantic.error_wrappers import ValidationError
-from pydantic.main import BaseModel
-from sqlalchemy.engine import create_engine
-from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.exc import DataError, OperationalError
 from sqlalchemy.orm.session import Session
-from sqlalchemy.sql.schema import Column
-from sqlalchemy.sql.sqltypes import Integer, String
 
 from crud_factory.factories import CRUDFactory
-
-Base = declarative_base()
-
-
-@pytest.fixture(
-    params=["sqlite:///", "postgres://postgres:postgres@localhost:5432/test"]
-)
-def engine(request):
-    return create_engine(request.param)
-
-
-@pytest.fixture(autouse=True)
-def setup_database(engine):
-    Base.metadata.create_all(engine)
-
-
-@pytest.fixture()
-def session(engine):
-    _session = Session(bind=engine)
-    yield _session
-    _session.close()
-
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True)
-    name = Column(String(30))
-    age = Column(Integer)
-
-
-class UserCreate(BaseModel):
-    name: str
-    age: int
-
-
-class UserUpdate(BaseModel):
-    name: str
-
-
-class UserOut(BaseModel):
-    id: int
-    name: str
-    age: int
-
-    class Config:
-        orm_mode = True
+from tests.utils import User, UserCreate, UserOut, UserUpdate
 
 
 def test_create_with_dict(session: Session):
@@ -75,8 +24,9 @@ def test_create_wrong_column_name(session: Session):
         crud_user.create(session, {"name": "Potato", "wrong": 7})
 
 
-# NOTE: This works on SQLite.
 def test_create_wrong_value_type(session: Session):
+    if session.bind.url.drivername in ("sqlite"):
+        pytest.skip()
     crud_user = CRUDFactory.get_sqlalchemy(User, UserCreate, UserUpdate)
-    obj = crud_user.create(session, {"name": "Potato", "age": "Potato"})
-    assert obj.age == "Potato" and obj.name == "Potato", UserOut.from_orm(obj)
+    with pytest.raises((DataError, OperationalError)):
+        crud_user.create(session, {"name": "Potato", "age": "Potato"})
